// lib/api/trip/trip.types.ts
import { TripServiceType } from "@/lib/shared/types/trip-service-type.type";

export interface CreateTripPayload {
  origin: string;
  destination: string;
  departure: string;
  arrival?: string;
  serviceType: TripServiceType;
  vehicleId?: string;
  capacity?: number;
  minPassengers?: number;
  autoGenerated?: boolean;
  basePrice: number;
}

export interface Trip {
  id: string;
  origin: string;
  destination: string;
  departure: string;
  arrival?: string;
  basePrice: number;
  serviceType: TripServiceType;
  vehicleId?: string;
  capacity?: number;
  minPassengers?: number;
  autoGenerated: boolean;
  createdAt: string;
  updatedAt: string;
}